final jir

+++ src/main/java/org/example/Elemen.java
@@ -6,7 +6,7 @@ public class Elemen {
     Elemen prev;
     data info;
     
-    Elemen(String nama, String penulis, String penerbit, int harga){
+    Elemen(String nama, String penulis, String penerbit, float harga){
         info = new data(nama, penulis, penerbit, harga);
         next = null;
         prev = null;
