+++ src/main/java/org/example/data.java
@@ -2,9 +2,9 @@ package org.example;
 
 public class data {
     String nama, penulis, penerbit;
-    int harga;
+    float harga;
 
-    public data(String nama, String penulis, String penerbit, int harga) {
+    public data(String nama, String penulis, String penerbit, float harga) {
         this.nama = nama;
         this.penulis = penulis;
         this.penerbit = penerbit;
